Add stdout_pipe/stderr_pipe parameters.
Instead of logging directly to a file, instead we redirect STDOUT/STDERR to a
pipe, which allows for redirection of the daemon's log output through an
arbitrary pipeline (perhaps e.g. immediately compressing the daemon's output to
disk using gzip) or feeding the output into a pipe-based log collector such as
cronolog.